Fixes in header
